@@ -29,15 +29,17 @@
         }
     ],
     "require":{
-        "php": ">=7.1.0",
-        "cakephp/cakephp": "^3.7",
-        "friendsofcake/crud": "^5.4.6",
-        "neomerx/json-api": "^3.0"
+        "php": ">=7.2.0",
+        "cakephp/cakephp": "^4.0.2",
+        "friendsofcake/crud": "^6.0-beta",
+        "neomerx/json-api": "^4.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7.14|^6.0",
-        "friendsofcake/cakephp-test-utilities": "dev-master",
-        "friendsofcake/search": "^4.4"
+        "phpunit/phpunit": "~8.5",
+        "friendsofcake/cakephp-test-utilities": "^2.0.1",
+        "friendsofcake/search": "^6.0.0",
+        "cakephp/cakephp-codesniffer": "^4.0",
+        "dms/phpunit-arraysubset-asserts": "^0.1.0"
     },
     "autoload": {
         "psr-4": {
@@ -47,6 +49,7 @@
     "autoload-dev": {
         "psr-4": {
             "CrudJsonApi\\Test\\": "tests",
+            "CrudJsonApi\\Test\\App\\": "tests/test_app/src",
             "Cake\\Test\\Fixture\\": "./vendor/cakephp/cakephp/tests/Fixture"
         }
     },
@@ -58,5 +61,12 @@
         "issues":"https://github.com/FriendsOfCake/crud-json-api/issues",
         "wiki":"http://crud.readthedocs.org/en/latest/",
         "irc":"irc://irc.freenode.org/friendsofcake"
+    },
+    "scripts": {
+        "cs-check": "phpcs -p --standard=vendor/cakephp/cakephp-codesniffer/CakePHP src/ tests/",
+        "cs-fix": "phpcbf --standard=vendor/cakephp/cakephp-codesniffer/CakePHP src/ tests/",
+        "phpstan-setup": "cp composer.json composer.backup && composer require --dev phpstan/phpstan:\"^0.12\" && mv composer.backup composer.json",
+        "phpstan": "phpstan analyse src/",
+        "test": "phpunit"
     }
 }
