@@ -1,13 +1,16 @@
 <?php
+declare(strict_types=1);
+
 namespace CrudJsonApi\Test\TestCase\Schema\JsonApi;
 
 use Cake\Controller\Controller;
 use Cake\ORM\TableRegistry;
 use Cake\View\View;
+use Crud\TestSuite\TestCase;
 use CrudJsonApi\Listener\JsonApiListener;
 use CrudJsonApi\Schema\JsonApi\DynamicEntitySchema;
-use Crud\TestSuite\TestCase;
 use Neomerx\JsonApi\Contracts\Factories\FactoryInterface;
+use Neomerx\JsonApi\Contracts\Schema\ContextInterface;
 use Neomerx\JsonApi\Contracts\Schema\SchemaInterface;
 use Neomerx\JsonApi\Factories\Factory;
 use Neomerx\JsonApi\Schema\Identifier;
@@ -18,7 +21,6 @@ use Neomerx\JsonApi\Schema\Identifier;
  */
 class DynamicEntitySchemaTest extends TestCase
 {
-
     /**
      * fixtures property
      *
@@ -31,6 +33,16 @@ class DynamicEntitySchemaTest extends TestCase
     ];
 
     /**
+     * @return void
+     */
+    public function setUp(): void
+    {
+        parent::setUp();
+
+        require CONFIG . 'routes.php';
+    }
+
+    /**
      * Test NeoMerx override getAttributes().
      *
      * @return void
@@ -41,7 +53,7 @@ class DynamicEntitySchemaTest extends TestCase
         $table = TableRegistry::get('Countries');
         $query = $table->find()
             ->where([
-                'Countries.id' => 2
+                'Countries.id' => 2,
             ])
             ->contain([
                 'Cultures',
@@ -72,12 +84,12 @@ class DynamicEntitySchemaTest extends TestCase
         // make view return associations on get('_associations') call
         $view = $this
             ->getMockBuilder(View::class)
-            ->setMethods(['get'])
+            ->onlyMethods(['get'])
             ->disableOriginalConstructor()
             ->getMock();
 
-        $view->set('_repositories', $repositories);
-        $view->set('_inflect', 'dasherize');
+        $view->setConfig('repositories', $repositories);
+        $view->setConfig('inflect', 'dasherize');
 
         // setup the schema
         $schemaFactoryInterface = $this
@@ -87,16 +99,20 @@ class DynamicEntitySchemaTest extends TestCase
 
         $schema = $this
             ->getMockBuilder(DynamicEntitySchema::class)
-            ->setMethods(null)
             ->setConstructorArgs([$schemaFactoryInterface, $view, $table])
+            ->onlyMethods([])
             ->getMock();
 
+        $this->setReflectionClassInstance($schema, DynamicEntitySchema::class);
         $this->setReflectionClassInstance($schema);
 
-        $this->setProtectedProperty('view', $view, $schema);
+        $this->setProtectedProperty('view', $view, DynamicEntitySchema::class);
+
+        $context = $this->getMockBuilder(ContextInterface::class)
+            ->getMock();
 
         // assert method
-        $result = $this->callProtectedMethod('getAttributes', [$entity], $schema);
+        $result = $this->callProtectedMethod('getAttributes', [$entity, $context], $schema);
 
         $this->assertSame('BG', $result['code']);
         $this->assertArrayNotHasKey('id', $result);
@@ -117,7 +133,7 @@ class DynamicEntitySchemaTest extends TestCase
         $table = TableRegistry::get('Countries');
         $query = $table->find()
             ->where([
-                'Countries.id' => 2
+                'Countries.id' => 2,
             ])
             ->contain([
                 'Cultures',
@@ -148,15 +164,17 @@ class DynamicEntitySchemaTest extends TestCase
         // make view return associations on get('_associations') call
         $view = new View();
 
-        $view->set('_repositories', $repositories);
-        $view->set('_absoluteLinks', false); // test relative links (listener default)
-        $view->set('_inflect', 'dasherize');
+        $view->setConfig('repositories', $repositories);
+        $view->setConfig('absoluteLinks', false); // test relative links (listener default)
+        $view->setConfig('inflect', 'dasherize');
 
         // setup the schema
         $schema = new DynamicEntitySchema(new Factory(), $view, $table);
 
         // assert getRelationships()
-        $relationships = $schema->getRelationships($entity);
+        $context = $this->getMockBuilder(ContextInterface::class)
+            ->getMock();
+        $relationships = $schema->getRelationships($entity, $context);
 
         $this->assertArrayHasKey('currency', $relationships);
         $this->assertSame($expectedCurrencyId, $relationships['currency'][SchemaInterface::RELATIONSHIP_DATA]['id']);
@@ -171,27 +189,27 @@ class DynamicEntitySchemaTest extends TestCase
         $expected = '/currencies/1';
         $result = $schema->getRelationshipSelfLink($entity, 'currency');
         $this->setReflectionClassInstance($result);
-        $this->assertSame($expected, $this->getProtectedProperty('value', $result));
+        $this->assertSame($expected, $this->getProtectedProperty('value', get_class($result)));
 
         // assert generated belongsToLink using JsonApi (indirect link, requires custom JsonApiRoute)
         $view->set('_jsonApiBelongsToLinks', true);
         $expected = '/countries/2/relationships/currency';
         $result = $schema->getRelationshipSelfLink($entity, 'currency');
         $this->setReflectionClassInstance($result);
-        $this->assertSame($expected, $this->getProtectedProperty('value', $result));
+        $this->assertSame($expected, $this->getProtectedProperty('value', get_class($result)));
 
         // assert _ getRelationshipSelfLinks() for plural (hasMany)
         $expected = '/cultures?country_id=2';
 
         $result = $schema->getRelationshipRelatedLink($entity, 'cultures');
         $this->setReflectionClassInstance($result);
-        $this->assertSame($expected, $this->getProtectedProperty('value', $result));
+        $this->assertSame($expected, $this->getProtectedProperty('value', get_class($result)));
 
         // assert N-1 (i.e. belongs-to)  relationships are always included as a relationship
         unset($entity['currency']);
         $this->assertArrayNotHasKey('currency', $entity);
 
-        $result = $schema->getRelationships($entity);
+        $result = $schema->getRelationships($entity, $context);
         $this->assertArrayHasKey('currency', $result);
         $this->assertArrayHasKey('cultures', $result);
         $this->assertInstanceOf(Identifier::class, $result['currency'][DynamicEntitySchema::RELATIONSHIP_DATA]);
@@ -201,7 +219,7 @@ class DynamicEntitySchemaTest extends TestCase
         $table = TableRegistry::get('Countries');
         $query = $table->find()
             ->where([
-                'Countries.id' => 2
+                'Countries.id' => 2,
             ])
             ->contain([
                 'Cultures',
@@ -212,14 +230,14 @@ class DynamicEntitySchemaTest extends TestCase
         unset($entity['cultures']);
         $this->assertArrayNotHasKey('cultures', $entity);
 
-        $result = $schema->getRelationships($entity);
+        $result = $schema->getRelationships($entity, $context);
         $this->assertArrayNotHasKey('cultures', $result);
         $this->assertArrayHasKey('currency', $result);
 
         // test custom foreign key
         $query = $table->find()
             ->where([
-                'Countries.id' => 4
+                'Countries.id' => 4,
             ])
             ->contain([
                 'SuperCountries',
@@ -232,7 +250,7 @@ class DynamicEntitySchemaTest extends TestCase
         // test custom propertyName
         $query = $table->find()
             ->where([
-                'Countries.id' => 3
+                'Countries.id' => 3,
             ])
             ->contain([
                 'SubCountries',
