@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 namespace CrudJsonApi\Schema\JsonApi;
 
 use Cake\Core\App;
@@ -8,7 +10,9 @@ use Cake\ORM\Table;
 use Cake\Routing\Router;
 use Cake\Utility\Inflector;
 use Cake\View\View;
+use InvalidArgumentException;
 use Neomerx\JsonApi\Contracts\Factories\FactoryInterface;
+use Neomerx\JsonApi\Contracts\Schema\ContextInterface;
 use Neomerx\JsonApi\Contracts\Schema\LinkInterface;
 use Neomerx\JsonApi\Schema\BaseSchema;
 use Neomerx\JsonApi\Schema\Identifier;
@@ -21,6 +25,7 @@ class DynamicEntitySchema extends BaseSchema
 {
     /**
      * Holds the instance of Cake\View\View
+     *
      * @var \Cake\View\View
      */
     protected $view;
@@ -32,9 +37,9 @@ class DynamicEntitySchema extends BaseSchema
     /**
      * Class constructor
      *
-     * @param \Neomerx\JsonApi\Contracts\Factories\FactoryInterface $factory ContainerInterface
-     * @param \Cake\View\View $view Instance of the cake view we are rendering this in
-     * @param \Cake\ORM\Table $repository Repository to use
+     * @param \Neomerx\JsonApi\Contracts\Factories\FactoryInterface $factory    ContainerInterface
+     * @param \Cake\View\View                                       $view       Instance of the cake view we are rendering this in
+     * @param \Cake\ORM\Table                                       $repository Repository to use
      */
     public function __construct(
         FactoryInterface $factory,
@@ -49,7 +54,6 @@ class DynamicEntitySchema extends BaseSchema
 
     /**
      * @param \Cake\ORM\Table $repository The repository object
-     *
      * @return mixed
      */
     private function getTypeFromRepository(Table $repository)
@@ -72,16 +76,22 @@ class DynamicEntitySchema extends BaseSchema
     /**
      * Get resource id.
      *
-     * @param \Cake\ORM\Entity $entity Entity
+     * @param  \Cake\ORM\Entity $entity Entity
      * @return string
      */
     public function getId($entity): ?string
     {
-        return (string)$entity->get($this->repository->getPrimaryKey());
+        $primaryKey = $this->repository->getPrimaryKey();
+
+        if (is_array($primaryKey)) {
+            throw new \RuntimeException('Crud-Json-Api does not support composite keys out of the box.');
+        }
+
+        return (string)$entity->get($primaryKey);
     }
 
     /**
-     * @param \Cake\Datasource\EntityInterface $entity Entity
+     * @param  \Cake\Datasource\EntityInterface $entity Entity
      * @return \Cake\ORM\Table
      */
     protected function getRepository($entity = null): Table
@@ -92,7 +102,7 @@ class DynamicEntitySchema extends BaseSchema
 
         $repositoryName = $entity->getSource();
 
-        return $this->view->get('_repositories')[$repositoryName];
+        return $this->view->getConfig('repositories')[$repositoryName];
     }
 
     /**
@@ -101,7 +111,7 @@ class DynamicEntitySchema extends BaseSchema
      *
      * This method will ignore any properties that are entities.
      *
-     * @param \Cake\Datasource\EntityInterface $entity Entity
+     * @param  \Cake\Datasource\EntityInterface $entity Entity
      * @return array
      */
     protected function entityToShallowArray(EntityInterface $entity)
@@ -136,9 +146,10 @@ class DynamicEntitySchema extends BaseSchema
      * as JsonApi `attributes`.
      *
      * @param \Cake\Datasource\EntityInterface $entity Entity
+     * @param \Neomerx\JsonApi\Contracts\Schema\ContextInterface $context The Context
      * @return array
      */
-    public function getAttributes($entity): iterable
+    public function getAttributes($entity, ContextInterface $context): iterable
     {
         $entity->setHidden((array)$this->getRepository()->getPrimaryKey(), true);
 
@@ -177,10 +188,11 @@ class DynamicEntitySchema extends BaseSchema
      *
      * JSON API optional `related` links not implemented yet.
      *
-     * @param \Cake\Datasource\EntityInterface $entity Entity object
+     * @param  \Cake\Datasource\EntityInterface $entity Entity object
+     * @param \Neomerx\JsonApi\Contracts\Schema\ContextInterface $context The Context
      * @return array
      */
-    public function getRelationships($entity): iterable
+    public function getRelationships($entity, ContextInterface $context): iterable
     {
         $relations = [];
 
@@ -211,7 +223,10 @@ class DynamicEntitySchema extends BaseSchema
             }
 
             if (!$data && !is_array($foreignKey)) {
-                $data = new Identifier($entity->get($foreignKey), $this->getTypeFromRepository($association->getTarget()));
+                $data = new Identifier(
+                    (string)$entity->get($foreignKey),
+                    $this->getTypeFromRepository($association->getTarget())
+                );
             }
 
             $isOne = \in_array($association->type(), [Association::MANY_TO_ONE, Association::ONE_TO_ONE]);
@@ -228,7 +243,7 @@ class DynamicEntitySchema extends BaseSchema
     /**
      * NeoMerx override used to generate `self` links
      *
-     * @param \Cake\ORM\Entity|null $entity Entity, null only to be compatible with the Neomerx method
+     * @param  \Cake\ORM\Entity|null $entity Entity, null only to be compatible with the Neomerx method
      * @return string
      */
     public function getSelfSubUrl($entity = null): string
@@ -239,15 +254,17 @@ class DynamicEntitySchema extends BaseSchema
 
         $keys = array_values($entity->extract((array)$this->getRepository()->getPrimaryKey()));
 
-        return Router::url($this->_getRepositoryRoutingParameters($this->repository) + $keys + [
+        return Router::url(
+            $this->_getRepositoryRoutingParameters($this->repository) + $keys + [
             '_method' => 'GET',
             'action' => 'view',
-        ], $this->view->get('_absoluteLinks'));
+            ],
+            $this->view->getConfig('absoluteLinks', false)
+        );
     }
 
     /**
      * @param string $name Relationship name in lowercase singular or plural
-     *
      * @return \Cake\ORM\Association|null
      */
     protected function getAssociationByProperty(string $name): ?Association
@@ -268,15 +285,14 @@ class DynamicEntitySchema extends BaseSchema
      * Example: /cultures?country_id=1 (or /country/1/cultures if your routes are configured like this)
      *
      * @param \Cake\Datasource\EntityInterface $entity Entity
-     * @param string $name Relationship name in lowercase singular or plural
-     *
+     * @param string                           $name   Relationship name in lowercase singular or plural
      * @return \Neomerx\JsonApi\Contracts\Schema\LinkInterface
      */
     public function getRelationshipSelfLink($entity, string $name): LinkInterface
     {
         $association = $this->getAssociationByProperty($name);
         if (!$association) {
-            throw new \InvalidArgumentException('Invalid association ' . $name);
+            throw new InvalidArgumentException('Invalid association ' . $name);
         }
 
         $relatedRepository = $association->getTarget();
@@ -286,13 +302,16 @@ class DynamicEntitySchema extends BaseSchema
             [, $controllerName] = pluginSplit($this->getRepository()->getRegistryAlias());
             $sourceName = Inflector::underscore(Inflector::singularize($controllerName));
 
-            $url = Router::url($this->_getRepositoryRoutingParameters($relatedRepository) + [
+            $url = Router::url(
+                $this->_getRepositoryRoutingParameters($relatedRepository) + [
                 '_method' => 'GET',
                 'action' => 'view',
                 $sourceName . '_id' => $entity->id,
                 'from' => $this->getRepository()->getRegistryAlias(),
                 'type' => $name,
-            ], $this->view->get('_absoluteLinks'));
+                ],
+                $this->view->getConfig('absoluteLinks', false)
+            );
         } else {
             $name = Inflector::dasherize($name);
             $relatedEntity = $entity->get($name);
@@ -303,10 +322,13 @@ class DynamicEntitySchema extends BaseSchema
                 $keys = array_values($entity->extract((array)$association->getForeignKey()));
             }
 
-            $url = Router::url($this->_getRepositoryRoutingParameters($relatedRepository) + $keys + [
+            $url = Router::url(
+                $this->_getRepositoryRoutingParameters($relatedRepository) + $keys + [
                 '_method' => 'GET',
                 'action' => 'view',
-            ], $this->view->get('_absoluteLinks'));
+                ],
+                $this->view->getConfig('absoluteLinks', false)
+            );
         }
 
         return $this->getFactory()->createLink(false, $url, false);
@@ -319,15 +341,14 @@ class DynamicEntitySchema extends BaseSchema
      * hasMany example"   /countries/1/currencies"
      *
      * @param \Cake\Datasource\EntityInterface $entity Entity
-     * @param string $name Relationship name in lowercase singular or plural
-     *
+     * @param string                           $name   Relationship name in lowercase singular or plural
      * @return \Neomerx\JsonApi\Contracts\Schema\LinkInterface
      */
     public function getRelationshipRelatedLink($entity, string $name): LinkInterface
     {
         $association = $this->getAssociationByProperty($name);
         if (!$association) {
-            throw new \InvalidArgumentException('Invalid association ' . $name);
+            throw new InvalidArgumentException('Invalid association ' . $name);
         }
 
         $relatedRepository = $association->getTarget();
@@ -337,10 +358,14 @@ class DynamicEntitySchema extends BaseSchema
         $primaryKeys = $entity->extract((array)$this->getRepository()->getPrimaryKey());
         $keys = array_combine($foreignKeys, $primaryKeys);
 
-        $url = Router::url($this->_getRepositoryRoutingParameters($relatedRepository) + $keys + [
+        $url = Router::url(
+            $this->_getRepositoryRoutingParameters($relatedRepository) + $keys + [
                 '_method' => 'GET',
                 'action' => 'index',
-            ], $this->view->get('_absoluteLinks'));
+                '?' => $keys
+            ],
+            $this->view->getConfig('absoluteLinks', false)
+        );
 
         return $this->getFactory()
             ->createLink(false, $url, false);
@@ -350,7 +375,7 @@ class DynamicEntitySchema extends BaseSchema
      * Parses the name of an Entity class to build a lowercase plural
      * controller name to be used in links.
      *
-     * @param \Cake\Datasource\RepositoryInterface $repository Repository
+     * @param  \Cake\Datasource\RepositoryInterface $repository Repository
      * @return array Array holding lowercase controller name as the value
      */
     protected function _getRepositoryRoutingParameters($repository)
@@ -360,7 +385,7 @@ class DynamicEntitySchema extends BaseSchema
 
         return [
             'controller' => $controllerName,
-            'plugin' => $pluginName
+            'plugin' => $pluginName,
         ];
     }
 }
