@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 namespace CrudJsonApi\Test\TestCase\Integration\JsonApi;
 
 use Cake\Core\Configure;
@@ -17,7 +19,7 @@ class ErrorsIntegrationTest extends JsonApiBaseTestCase
         $this->assertResponseCode(404);
         $this->_assertJsonApiResponseHeaders();
 
-        $actualResponseBody = $this->_getResponseWithEmptyDebugNode($this->_getBodyAsString());
+        $actualResponseBody = $this->_getResponseWithEmptyDebugNode($this->_getBodyAsString() . "\n");
         $this->assertResponseSameAsFile('Errors' . DS . '404-error-for-collection-in-debug-mode.json', $actualResponseBody);
     }
 
