@@ -1,8 +1,12 @@
 <?php
+declare(strict_types=1);
+
 namespace CrudJsonApi\Test\TestCase\Listener\JsonApi;
 
-use CrudJsonApi\Listener\JsonApi\DocumentValidator;
+use Cake\ORM\Entity;
+use Crud\Error\Exception\ValidationException;
 use Crud\TestSuite\TestCase;
+use CrudJsonApi\Listener\JsonApi\DocumentValidator;
 use Neomerx\JsonApi\Contracts\Schema\LinkInterface;
 use Neomerx\JsonApi\Schema\ErrorCollection;
 use stdClass;
@@ -22,22 +26,22 @@ class DocumentValidatorTest extends TestCase
      * Create a DocumentValidator instance for every test with disabled
      * aboutLinks config option.
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
         $this->_validator = new DocumentValidator([], []);
         $this->setReflectionClassInstance($this->_validator);
 
         $listenerConfig = [
-            'docValidatorAboutLinks' => false
+            'docValidatorAboutLinks' => false,
         ];
-        $this->setProtectedProperty('_config', $listenerConfig, $this->_validator);
+        $this->setProtectedProperty('_config', $listenerConfig, DocumentValidator::class);
     }
 
     /**
      * tearDown.
      */
-    public function tearDown()
+    public function tearDown(): void
     {
         parent::tearDown();
         unset($this->_validator);
@@ -45,70 +49,67 @@ class DocumentValidatorTest extends TestCase
 
     /**
      * _validateCreateDocument()
-     *
-     * @expectedException \Crud\Error\Exception\ValidationException
-     * @expectedExceptionMessage A validation error occurred
      */
     public function testValidateCreateDocument()
     {
+        $this->expectException(ValidationException::class);
+        $this->expectExceptionMessage('A validation error occurred');
         // assert success
         $document = [
             'data' => [
                 'type' => 'must-be-string',
-                'id' => 'd0b31ee1-4637-48c9-b9ef-fcefbb83d86f'
-            ]
+                'id' => 'd0b31ee1-4637-48c9-b9ef-fcefbb83d86f',
+            ],
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
-        $this->assertNull($this->_validator->validateCreateDocument());
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
+        $this->_validator->validateCreateDocument();
 
         // assert exception
         $document['data']['id'] = 'not-a-valid-uuid';
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->_validator->validateCreateDocument();
     }
 
     /**
      * _validateUpdateDocument()
-     *
-     * @expectedException \Crud\Error\Exception\ValidationException
-     * @expectedExceptionMessage A validation error occurred
      */
     public function testValidateUpdateDocument()
     {
+        $this->expectException('Crud\Error\Exception\ValidationException');
+        $this->expectExceptionMessage('A validation error occurred');
         // assert success
         $document = [
             'data' => [
                 'type' => 'must-be-string',
                 'id' => 'must-be-string',
-            ]
+            ],
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertNull($this->_validator->validateUpdateDocument());
 
         // assert exception
         $document['data']['id'] = null;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->_validator->validateUpdateDocument();
     }
 
     /**
      * _documentMustHavePrimaryData()
-     *
-     * @expectedException \Crud\Error\Exception\ValidationException
-     * @expectedExceptionMessage A validation error occurred
      */
     public function testDocumentMustHavePrimaryData()
     {
+        $this->expectException('Crud\Error\Exception\ValidationException');
+        $this->expectExceptionMessage('A validation error occurred');
         // assert success
         $document = [
-            'data' => []
+            'data' => [],
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
-        $this->assertTrue($this->callProtectedMethod('_documentMustHavePrimaryData', [], $this->_validator));
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
+        $this->assertTrue($this->callProtectedMethod('_documentMustHavePrimaryData', [], DocumentValidator::class));
 
         // assert exception
         $document = [];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->callProtectedMethod('_documentMustHavePrimaryData', [], $this->_validator);
     }
 
@@ -122,20 +123,20 @@ class DocumentValidatorTest extends TestCase
         // assert success
         $document = [
             'data' => [
-                'type' => 'must-be-string'
-            ]
+                'type' => 'must-be-string',
+            ],
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMustHaveType', [], $this->_validator));
 
         // assert false for non-string
         $document['data']['type'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_primaryDataMustHaveType', [], $this->_validator));
 
         // assert exception
         $document = [];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->callProtectedMethod('_primaryDataMustHaveType', [], $this->_validator);
     }
 
@@ -149,20 +150,20 @@ class DocumentValidatorTest extends TestCase
         // assert success
         $document = [
             'data' => [
-                'id' => 'edd28c99-216b-4ef7-a806-d865aca14f17'
-            ]
+                'id' => 'edd28c99-216b-4ef7-a806-d865aca14f17',
+            ],
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMayHaveUuid', [], $this->_validator));
 
         // assert false for non-string
         $document['data']['id'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_primaryDataMayHaveUuid', [], $this->_validator));
 
         // assert exception
         $document = [];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->callProtectedMethod('_primaryDataMayHaveUuid', [], $this->_validator);
     }
 
@@ -178,19 +179,19 @@ class DocumentValidatorTest extends TestCase
             'data' => [
                 'type' => 'must-be-string',
                 'id' => 'must-be-string',
-            ]
+            ],
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMustHaveId', [], $this->_validator));
 
         // assert false for non-string
         $document['data']['id'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_primaryDataMustHaveId', [], $this->_validator));
 
         // assert exception
         $document = [];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->callProtectedMethod('_primaryDataMustHaveId', [], $this->_validator);
     }
 
@@ -210,51 +211,51 @@ class DocumentValidatorTest extends TestCase
                         'data' => [
                             [
                                 'type' => 'cultures',
-                                'id' => '1'
-                            ]
-                        ]
+                                'id' => '1',
+                            ],
+                        ],
                     ],
                     'currency' => [
                         'data' => [
                             'type' => 'currencies',
-                            'id' => '2'
-                        ]
-                    ]
-                ]
-            ]
+                            'id' => '2',
+                        ],
+                    ],
+                ],
+            ],
         ];
 
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMayHaveRelationships', [], $this->_validator));
 
         // assert foreach loop continues if belongsTo relation has empty/null data node
         $document['data']['relationships']['currency']['data'] = null;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMayHaveRelationships', [], $this->_validator));
 
         // assert foreach loop continues if belongsTo relation does not have data node
         unset($document['data']['relationships']['currency']['data']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMayHaveRelationships', [], $this->_validator));
 
         // assert foreach loop continues if hasMany relation has empty/null data node
         $document['data']['relationships']['cultures']['data'] = null;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMayHaveRelationships', [], $this->_validator));
 
         // assert foreach loop continues if hasMany relation does not have data node
         unset($document['data']['relationships']['cultures']['data']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMayHaveRelationships', [], $this->_validator));
 
         // assert non-pass if `relationships` exists but has no members
         $document['data']['relationships'] = [];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_primaryDataMayHaveRelationships', [], $this->_validator));
 
         // assert pass if `relationships` does not exist at all
         unset($document['data']['relationships']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_primaryDataMayHaveRelationships', [], $this->_validator));
     }
 
@@ -273,36 +274,36 @@ class DocumentValidatorTest extends TestCase
                         'data' => [
                             [
                                 'type' => 'cultures',
-                                'id' => '1'
-                            ]
-                        ]
+                                'id' => '1',
+                            ],
+                        ],
                     ],
                     'currency' => [
                         'data' => [
                             'type' => 'currencies',
-                            'id' => '2'
-                        ]
-                    ]
-                ]
-            ]
+                            'id' => '2',
+                        ],
+                    ],
+                ],
+            ],
         ];
 
         // assert success when belongsTo relation has data node
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_relationshipMustHaveData', ['data.relationships.currency'], $this->_validator));
 
         // assert fail when belongsTo relation is missing data node
         unset($document['data']['relationships']['currency']['data']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipMustHaveData', ['data.relationships.currency'], $this->_validator));
 
         // assert success when hasMany relation has data node
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_relationshipMustHaveData', ['data.relationships.cultures'], $this->_validator));
 
         // assert fail when hasMany relation is missing data node
         unset($document['data']['relationships']['cultures']['data']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipMustHaveData', ['data.relationships.cultures'], $this->_validator));
     }
 
@@ -321,46 +322,46 @@ class DocumentValidatorTest extends TestCase
                         'data' => [
                             [
                                 'type' => 'cultures',
-                                'id' => '1'
-                            ]
-                        ]
+                                'id' => '1',
+                            ],
+                        ],
                     ],
                     'currency' => [
                         'data' => [
                             'type' => 'currencies',
-                            'id' => '2'
-                        ]
-                    ]
-                ]
-            ]
+                            'id' => '2',
+                        ],
+                    ],
+                ],
+            ],
         ];
 
         // assert success when belongsTo relationship data has valid type node
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_relationshipDataMustHaveType', ['currency', 'data.relationships.currency'], $this->_validator));
 
         // assert fail when belongsTo relationship data has type node but it's value is not a string
         $document['data']['relationships']['currency']['data']['type'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipDataMustHaveType', ['currency', 'data.relationships.currency'], $this->_validator));
 
         // assert fail when belongsTo relationship data does not have type node
         unset($document['data']['relationships']['currency']['data']['type']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipDataMustHaveType', ['currency', 'data.relationships.currency'], $this->_validator));
 
         // assert success when all hasMany relationships have data with valid type node
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_relationshipDataMustHaveType', ['cultures', 'data.relationships.cultures.data.0'], $this->_validator));
 
         // assert fail when one of the hasMany relationships has data with type node but it's value is not a string
         $document['data']['relationships']['cultures']['data'][0]['type'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipDataMustHaveType', ['cultures', 'data.relationships.cultures.data.0'], $this->_validator));
 
         // assert fail when one of the hasMany relationships has data without type node
         unset($document['data']['relationships']['cultures']['data'][0]['type']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipDataMustHaveType', ['cultures', 'data.relationships.cultures.data.0'], $this->_validator));
     }
 
@@ -379,98 +380,98 @@ class DocumentValidatorTest extends TestCase
                         'data' => [
                             [
                                 'type' => 'cultures',
-                                'id' => '1'
-                            ]
-                        ]
+                                'id' => '1',
+                            ],
+                        ],
                     ],
                     'currency' => [
                         'data' => [
                             'type' => 'currencies',
-                            'id' => '2'
-                        ]
-                    ]
-                ]
-            ]
+                            'id' => '2',
+                        ],
+                    ],
+                ],
+            ],
         ];
 
         // assert success when belongsTo relationship data has valid id node
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_relationshipDataMustHaveId', ['currency', 'data.relationships.currency'], $this->_validator));
 
         // assert fail when belongsTo relationship data has id node but it's value is not a string
         $document['data']['relationships']['currency']['data']['id'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipDataMustHaveId', ['currency', 'data.relationships.currency'], $this->_validator));
 
         // assert fail when belongsTo relationship data does not have id node
         unset($document['data']['relationships']['currency']['data']['id']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipDataMustHaveId', ['currency', 'data.relationships.currency'], $this->_validator));
 
         // assert success when all hasMany relationships have data with valid id node
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_relationshipDataMustHaveId', ['cultures', 'data.relationships.cultures.data.0'], $this->_validator));
 
         // assert fail when one of the hasMany relationships has data with id node but it's value is not a string
         $document['data']['relationships']['cultures']['data'][0]['id'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipDataMustHaveId', ['cultures', 'data.relationships.cultures.data.0'], $this->_validator));
 
         // assert fail when one of the hasMany relationships has data without id node
         unset($document['data']['relationships']['cultures']['data'][0]['id']);
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_relationshipDataMustHaveId', ['cultures', 'data.relationships.cultures.data.0'], $this->_validator));
     }
 
     /**
      * _isString()
      *
-     * @expectedException \Crud\Error\Exception\CrudException
-     * @expectedExceptionMessage Document member 'dummy.path' does not exist
      * @return void
      */
     public function testIsString()
     {
+        $this->expectException('Crud\Error\Exception\CrudException');
+        $this->expectExceptionMessage('Document member \'dummy.path\' does not exist');
         // assert success
         $document = [
-            'data' => 'some-string'
+            'data' => 'some-string',
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_isString', ['data'], $this->_validator));
 
         // assert false for non-string
         $document['data'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_isString', ['data'], $this->_validator));
 
         // assert exception
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_isString', ['dummy.path'], $this->_validator));
     }
 
     /**
      * _isString()
      *
-     * @expectedException \Crud\Error\Exception\CrudException
-     * @expectedExceptionMessage Document member 'dummy.path' does not exist
      * @return void
      */
     public function testIsUuid()
     {
+        $this->expectException('Crud\Error\Exception\CrudException');
+        $this->expectExceptionMessage('Document member \'dummy.path\' does not exist');
         // assert success
         $document = [
-            'data' => 'edd28c99-216b-4ef7-a806-d865aca14f17'
+            'data' => 'edd28c99-216b-4ef7-a806-d865aca14f17',
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertTrue($this->callProtectedMethod('_isUuid', ['data'], $this->_validator));
 
         // assert false for non-string
         $document['data'] = 123;
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_isUuid', ['data'], $this->_validator));
 
         // assert exception
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
         $this->assertFalse($this->callProtectedMethod('_isUuid', ['dummy.path'], $this->_validator));
     }
 
@@ -489,19 +490,19 @@ class DocumentValidatorTest extends TestCase
                     'name' => 'bravo-kernel',
                     'test-null' => null,
                     'test-true' => true,
-                    'test-false' => false
+                    'test-false' => false,
                 ],
                 'sub-array' => [
                     [
                         'name' => 'bravo-kernel',
                         'test-null' => null,
                         'test-true' => true,
-                        'test-false' => false
-                    ]
-                ]
-            ]
+                        'test-false' => false,
+                    ],
+                ],
+            ],
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
 
         $this->assertTrue($this->callProtectedMethod('_hasProperty', ['data'], $this->_validator));
         $this->assertTrue($this->callProtectedMethod('_hasProperty', ['data.attributes'], $this->_validator));
@@ -523,26 +524,25 @@ class DocumentValidatorTest extends TestCase
 
     /**
      * _getProperty()
-     *
-     * @expectedException \Crud\Error\Exception\CrudException
-     * @expectedExceptionMessage Error retrieving a value for non-existing JSON API document property 'path.does.not.exist'
      */
     public function testGetProperty()
     {
+        $this->expectException('Crud\Error\Exception\CrudException');
+        $this->expectExceptionMessage('Error retrieving a value for non-existing JSON API document property \'path.does.not.exist\'');
         $this->_validator = new DocumentValidator([], []);
         $this->setReflectionClassInstance($this->_validator);
         $document = [
             'data' => [
                 'attributes' => [
-                    'name' => 'bravo-kernel'
-                ]
-            ]
+                    'name' => 'bravo-kernel',
+                ],
+            ],
         ];
-        $this->setProtectedProperty('_document', $document, $this->_validator);
+        $this->setProtectedProperty('_document', $document, DocumentValidator::class);
 
         //assert test single and multi-level lookups
         $expected = [
-           'name' => 'bravo-kernel'
+           'name' => 'bravo-kernel',
         ];
         $this->assertEquals($expected, $this->callProtectedMethod('_getProperty', ['data.attributes'], $this->_validator));
         $this->assertEquals('bravo-kernel', $this->callProtectedMethod('_getProperty', ['data.attributes.name'], $this->_validator));
@@ -616,18 +616,18 @@ class DocumentValidatorTest extends TestCase
 
         // assert no link is created if disabled in listener config
         $listenerConfig = [
-            'docValidatorAboutLinks' => false
+            'docValidatorAboutLinks' => false,
         ];
 
-        $this->setProtectedProperty('_config', $listenerConfig, $this->_validator);
+        $this->setProtectedProperty('_config', $listenerConfig, DocumentValidator::class);
         $this->assertNull($this->callProtectedMethod('_getAboutLink', ['http://www.friendsofcake.com'], $this->_validator));
 
         // assert link is created when enabled in listener config
         $listenerConfig = [
-            'docValidatorAboutLinks' => true
+            'docValidatorAboutLinks' => true,
         ];
 
-        $this->setProtectedProperty('_config', $listenerConfig, $this->_validator);
+        $this->setProtectedProperty('_config', $listenerConfig, DocumentValidator::class);
         $result = $this->callProtectedMethod('_getAboutLink', ['http://www.friendsofcake.com'], $this->_validator);
 
         $this->assertInstanceOf(LinkInterface::class, $result);
@@ -645,7 +645,7 @@ class DocumentValidatorTest extends TestCase
         $this->setReflectionClassInstance($this->_validator);
 
         $result = $this->callProtectedMethod('_getErrorCollectionEntity', [], $this->_validator);
-        $this->assertInstanceOf('\Cake\ORM\Entity', $result);
+        $this->assertInstanceOf(Entity::class, $result);
 
         $errors = $result->getErrors();
         $this->assertArrayHasKey('CrudJsonApiListener', $errors);
