@@ -1,18 +1,26 @@
 <?php
+declare(strict_types=1);
+
 namespace CrudJsonApi\Test\App;
 
 use Cake\Http\BaseApplication;
+use Cake\Http\Middleware\BodyParserMiddleware;
+use Cake\Http\MiddlewareQueue;
 use Cake\Routing\Middleware\AssetMiddleware;
 use Cake\Routing\Middleware\RoutingMiddleware;
 
 class Application extends BaseApplication
 {
-
-    public function middleware($middlewareQueue)
+    public function middleware(MiddlewareQueue $middlewareQueue): MiddlewareQueue
     {
+        $bodies = new BodyParserMiddleware();
+        $bodies->addParser(['application/vnd.api+json'], function ($body) {
+            return json_decode($body, true);
+        });
         $middlewareQueue
             // Handle plugin/theme assets like CakePHP normally does.
             ->add(AssetMiddleware::class)
+            ->add($bodies)
 
             // Add routing middleware.
             ->add(new RoutingMiddleware($this));
