@@ -1,21 +1,24 @@
 <?php
+declare(strict_types=1);
+
 namespace CrudJsonApi\Test\App\Model\Table;
 
+use Cake\ORM\Table;
 use Cake\Validation\Validator;
 
-class NationalCitiesTable extends \Cake\ORM\Table
+class NationalCitiesTable extends Table
 {
-    public function initialize(array $config)
+    public function initialize(array $config): void
     {
         $this->belongsTo('Countries');
     }
 
     // prevent unintentionally creating hasMany records
-    public function validationDefault(Validator $validator)
+    public function validationDefault(Validator $validator): Validator
     {
         $validator
             ->requirePresence('name', 'create')
-            ->notEmpty('name');
+            ->notEmptyString('name');
 
         return $validator;
     }
